services/zoe-data/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.5
aiosqlite==0.20.0
python-multipart==0.0.12
httpx==0.28.0
websockets==14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
ag-ui-protocol==0.1.14
